requirements.txt
openai>=1.60
python-dotenv
Pillow
pydantic>=2
